+++ src/Grafana.OpenTelemetry.Base/ResourceDetectors/AWSEBSDetectorInitializer.cs
@@ -1,0 +1,21 @@
+//
+// Copyright Grafana Labs
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#if !NETSTANDARD
+using OpenTelemetry.Resources;
+
+namespace Grafana.OpenTelemetry
+{
+    internal class AWSEBSDetectorInitializer : ResourceDetectorInitializer
+    {
+        public override ResourceDetector Id { get; } = ResourceDetector.AWSEBSDetector;
+
+        protected override ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder.AddAWSEBSDetector();
+        }
+    }
+}
+#endif