+++ src/Grafana.OpenTelemetry.Base/ResourceDetectors/AzureVMDetectorInitializer.cs
@@ -1,0 +1,19 @@
+//
+// Copyright Grafana Labs
+// SPDX-License-Identifier: Apache-2.0
+//
+
+using OpenTelemetry.Resources;
+
+namespace Grafana.OpenTelemetry
+{
+    internal class AzureVMDetectorInitializer : ResourceDetectorInitializer
+    {
+        public override ResourceDetector Id { get; } = ResourceDetector.AzureVMDetector;
+
+        protected override ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder.AddAzureVMDetector();
+        }
+    }
+}