+++ src/Grafana.OpenTelemetry.Base/ResourceDetectors/ContainerResource.cs
@@ -1,0 +1,23 @@
+//
+// Copyright Grafana Labs
+// SPDX-License-Identifier: Apache-2.0
+//
+
+using OpenTelemetry.Resources;
+
+#if NET6_0_OR_GREATER
+
+namespace Grafana.OpenTelemetry
+{
+    internal class ContainerResourceInitializer : ResourceDetectorInitializer
+    {
+        public override ResourceDetector Id { get; } = ResourceDetector.Container;
+
+        protected override ResourceBuilder InitializeResourceDetector(ResourceBuilder builder)
+        {
+            return builder.AddContainerDetector();
+        }
+    }
+}
+
+#endif